Renamed and moved to Common namespace
